requirements.txt
python-dotenv
fastapi
uvicorn
customtkinter
pydantic
python-jose
requests
bcrypt
pytest